+++ src/test/java/com/jeanchampemont/gunip/AbstractGenericUnitParserTest.java
@@ -10,10 +10,10 @@ public abstract class AbstractGenericUnitParserTest {
         super();
     }
 
-    protected void runTestCase(String input, double expectedTimeInS) throws GenericUnitParserException {
+    protected void runTestCase(String input, double expectedValueInBaseUnit) throws GenericUnitParserException {
         double parsedResult = 0;
         parsedResult = sut.parse(input);
-        Assert.assertEquals(expectedTimeInS, parsedResult, 0);
+        Assert.assertEquals(expectedValueInBaseUnit, parsedResult, 0);
     }
 
 }